requirements.txt
aniso8601==1.2.0
argcomplete==1.11.1
asn1crypto==1.3.0
awscli==1.18.44
beautifulsoup4==4.9.0
boto3==1.12.44
botocore==1.15.44
bs4==0.0.1
certifi==2020.4.5.1
cffi==1.14.0
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.1
colorama==0.4.3
country-converter==0.6.7
cryptography==2.1.4
docutils==0.15.2
durationpy==0.5
Flask==0.12.1
git+https://github.com/johnwheeler/flask-ask.git@v0.8.8
future==0.18.2
gunicorn==20.0.4
hjson==3.0.1
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
jieba==0.42.1
Jinja2==2.11.2
jmespath==0.9.5
kappa==0.6.0
MarkupSafe==1.1.1
mysql-connector==2.2.9
news-api==1.1.0
newsapi-python==0.2.6
numpy==1.18.3
pandas==1.0.3
pip-tools==5.0.0
placebo==0.9.0
pyasn1==0.4.8
pycountry==19.8.18
pycparser==2.20
PyMySQL==0.9.3
pyOpenSSL==17.0.0
python-dateutil==2.6.1
python-slugify==4.0.0
pytz==2019.3
PyYAML==3.12
requests==2.23.0
rsa==3.4.2
s3transfer==0.3.3
selenium==3.141.0
six==1.11.0
snownlp==0.12.3
soupsieve==2.0
text-unidecode==1.3
toml==0.10.0
tqdm==4.45.0
troposphere==2.6.0
true-encoding==1.0.2
urllib3==1.25.9
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.51.0
zipp==3.1.0
